+++ CHANGELOG.rst
@@ -10,10 +10,6 @@ dev
 - Added new module ``mass_definitions`` which more robustly deals with various halo mass definitions, and also includes
   ability to convert mass functions between different definitions.
 
-**Enhancement**
-
-- Added `get_dependencies` method to _Framework, to enable finding all parameters that a quantity depends on.
-
 **Bugfixes**
 
 - When using camb for the transfer function, some cosmologies would lead to a segfault (i.e. when Ob0 or Tcmb0 are not
@@ -24,6 +20,20 @@ dev
 - Removed logging, which was redundant.
 - Moved from nose to pytest
 - Significant overhaul of package structure to more modularised form.
+
+
+v3.0.8 [26th March 2020]
+------------------------
+**Bugfixes**
+
+- Some documentation and CI fixes.
+
+**Enhancement**
+
+- Added `get_dependencies` method to _Framework, to enable finding all parameters that a quantity depends on.
+- Growth factor speed-up (thanks to @jlashner)
+- Added use_splined_growth argument (thanks to @jlashner)
+
 
 v3.0.7 [29th April 2019]
 ------------------------